Load partial encoded tiles from broken tiff files
When not enough data is available in a scanline, load at least the
previous scanlines and show them instead of skipping the whole tile.
Set the remaining part of the buffer that was not yet read to 0.

This way, one can at least see a partial image instead of failing
the load operation completely.