test/tiffcp-thumbnail.sh: skip test if tools aren't built
In Gentoo, we avoid building the tools for multilib (32-bit, x86) builds on
amd64/x86_64 because we only need the library to keep binary applications working.

This causes a test failure in just tiffcp-thumbnail.sh as the 'thumbnail'
binary isn't built. Skip it if unavailable as it's only a single test.

Fixes: https://gitlab.com/libtiff/libtiff/-/issues/421